move role-info out of GameAdmin column
